Changed all floating calculations to doubles. Error was the same (slightly worse) with respect to gregers numbers. This change will be reverted in a future commit. Writes more csv files to aid in testing if truncating data before fitting spline will minimze the error wrt Greger

+++ computetaylor.cpp
@@ -212,7 +212,7 @@ void ComputeTaylor::createTaylorPolynomial(){
 
         // Create monomial term;
         // TODO put in check if number of betas is not as expected
-        taylorPolynomial+=betas.at(i+1)*monomialVector<float>(monomialExponentsVector);
+        taylorPolynomial+=betas.at(i+1)*monomialVector<double>(monomialExponentsVector);
         monomialExponentsVector.clear();
 
     }
@@ -223,13 +223,13 @@ void ComputeTaylor::createTaylorPolynomial(){
 
 }
 
-std::vector<float> ComputeTaylor::applyModelToOrder(std::vector<std::vector<float>> &orderVector){
+std::vector<double> ComputeTaylor::applyModelToOrder(std::vector<std::vector<double>> &orderVector){
     // For now the storage container is a 2D vector like:
     // {x0,x1,x2....xn},
     // {y1,y2,y3....yn},
     // {o1,o2,o3....on}
-    float evaluatedCorrection;
-    std::vector<float> correctedOrderLine;
+    double evaluatedCorrection;
+    std::vector<double> correctedOrderLine;
     for (int i=0;i<orderVector.size();i++){
 
         // Calculate the correction and stuff vector
@@ -245,7 +245,7 @@ bool ComputeTaylor::displayTaylorPolynomial(){
     std::cout<<"The Taylor polynomial is: "<<taylorPolynomial;
 }
 
-bool ComputeTaylor::updateBeta(std::vector<float> beta){
+bool ComputeTaylor::updateBeta(std::vector<double> beta){
 
     // TODO There are better ways to create the final taylorPolynomial than
     // doing it this way
